requirements.txt
# Flask and web framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# AI/ML libraries
torch>=2.6.0
torchvision>=0.19.0
transformers>=4.40.0
accelerate>=1.10.0
huggingface-hub>=0.34.3

# Image processing
Pillow>=11.3.0
opencv-python>=4.12.0.88
numpy>=2.2.6
scikit-image>=0.25.2
scipy>=1.15.3

# Database - MySQL (Windows compatible)
SQLAlchemy>=2.0.23
PyMySQL>=1.1.0
cryptography>=3.4.8

# API Documentation
flasgger==0.9.7.1

# Serialization
marshmallow==4.0.1

# Additional utilities
requests>=2.32.4
python-dotenv>=1.0.0
PyYAML>=6.0.2
tqdm>=4.67.1

# Background removal (if needed)
rembg>=2.0.67

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0